@@ -1,7 +1,7 @@
 import {Injectable} from '@angular/core'
 import {ShoppingCartService} from '../restaurant-detail/shopping-cart/shopping-cart.service'
 import {CartItem} from '../restaurant-detail/shopping-cart/cart-item.model'
-import {Http, Headers, RequestOptions} from '@angular/http'
+import {HttpClient} from '@angular/common/http'
 import {Observable} from 'rxjs/Observable'
 import 'rxjs/add/operator/map'
 import {Order, OrderItem} from './order.model'
@@ -9,7 +9,7 @@ import {MEAT_API} from '../app.api'
 @Injectable()
 
 export class OrderService {
-  constructor(private cartService: ShoppingCartService, private http: Http){}
+  constructor(private cartService: ShoppingCartService, private http: HttpClient){}
 
   itemsValue(): number {
     return this.cartService.total()
@@ -34,12 +34,7 @@ export class OrderService {
     this.cartService.clear()
   }
   checkOrder(order: Order): Observable<string>{
-    const headers = new Headers()
-    headers.append('Content-Type', 'application/json')
-    return this.http.post(`${MEAT_API}/orders`,
-                        JSON.stringify(order),
-                        new RequestOptions({headers: headers}))
-                    .map(response=> response.json())
+    return this.http.post<Order>(`${MEAT_API}/orders`, order)
                     .map(order => order.id)
   }
 }
